IMPALA-13020 (part 1): Change thrift_rpc_max_message_size to int64_t
Thrift 0.16.0 introduced a max message size to protect
receivers against a malicious message allocating large
amounts of memory. That limit is a 32-bit signed integer,
so the max value is 2GB. Impala introduced the
thrift_rpc_max_message_size startup option to set that
for Impala's thrift servers.

There are times when Impala wants to send a message that
is larger than 2GB. In particular, the catalog-update
topic for the statestore can exceed 2GBs when there is
a lot of metadata loaded using the old v1 catalog. When
there is a 2GB max message size, the statestore can create
and send a >2GB message, but the impalads will reject
it. This can lead to impalads having stale metadata.

This switches to a patched Thrift that uses an int64_t
for the max message size for C++ code. It does not modify
the limit.

The MaxMessageSize error was being swallowed in TAcceptQueueServer.cpp,
so this fixes that location to always print MaxMessageSize
exceptions.

This is only patching the Thrift C++ library. It does not
patch the Thrift Java library. There are a few reasons for
that:
 - This specific issue involves C++ to C++ communication and
   will be solved by patching the C++ library.
 - C++ is easy to patch as it is built via the native-toolchain.
   There is no corresponding mechanism for patching our Java
   dependencies (though one could be developed).
 - Java modifications have implications for other dependencies
   like Hive which use Thrift to communicate with HMS.
For the Java code that uses max message size, this converts
the 64-bit value to 32-bit value by capping the value at
Integer.MAX_VALUE.

Testing:
 - Added enough tables to produce a >2GB catalog-topic and
   restarted an impalad with a higher limit specific. Without
   the patch, the catalog-topic update would be rejected by the
   impalad. With the patch, it succeeds.

Reviewed-on: http://gerrit.cloudera.org:8080/21367
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>
(cherry picked from commit 13df8239d82a61afc3196295a7878ca2ffe91873)

+++ be/src/rpc/thrift-util.h
@@ -39,12 +39,12 @@ class TNetworkAddress;
 class ThriftServer;
 
 /// Default max message size from Thrift library.
-inline int ThriftDefaultMaxMessageSize() {
+inline int64_t ThriftDefaultMaxMessageSize() {
   return apache::thrift::TConfiguration::DEFAULT_MAX_MESSAGE_SIZE;
 }
 
 /// Return the effective max message size based on 'thrift_rpc_max_message_size' flag.
-int ThriftRpcMaxMessageSize();
+int64_t ThriftRpcMaxMessageSize();
 
 /// Return the default Thrift's TConfiguration based on given backend config flags.
 std::shared_ptr<apache::thrift::TConfiguration> DefaultTConfiguration();
